LFS puede recivir request y tiene una consola para pasarle request manual en simultaneo (implementacion de threads)

+++ LissandraFileSystem/recibir.h
@@ -29,7 +29,7 @@ typedef struct t_paquete_select
 {
 	char* nombre_tabla;
 	uint32_t nombre_tabla_long;
-	int valor_key;
+	uint16_t valor_key;
 }t_paquete_select;
 
 typedef struct t_paquete_insert
